ci(release): add new labels for features in release notes (#11)
* ci(release): add new labels for features in release notes

This commit updates the `release.yml` workflow to include `enhancement` and `new` labels in the "Features" section of the automatically generated release notes.

Signed-off-by: James Rich <2199651+jamesarich@users.noreply.github.com>

* chore(release): externalize changelog configuration

Moves the changelog generation configuration from the release workflow file into a separate `changelog-config.json` file. This improves the organization and maintainability of the release process.

The `release.yml` workflow was updated to reference this new configuration file and set additional parameters for the changelog builder action, including `owner`, `repository`, `ignorePreReleases`, `failOnError`, and `fetchViaCommits`.

Signed-off-by: James Rich <2199651+jamesarich@users.noreply.github.com>

---------

Signed-off-by: James Rich <2199651+jamesarich@users.noreply.github.com>